Refactor TrackList parsing
