+++ src/OszkConnector/Models/MekConvert.cs
@@ -61,71 +61,5 @@ namespace OszkConnector.Models
             }
             return cc;
         }
-
-        public static AudioBookTrack ConvertAudioTrackFrom(string text)
-        {
-            //Eg: "01_bojgas.mp3 - Itt kezdődik (10:53 min. 7,8 Mbyte)"
-            //     2     1    3           4                   5
-
-            var converterRegex = new Regex(@"((\d+)?.+\.(mp3))?\s?-\s?(.+)\s?(\(.+\))");
-            var match = converterRegex.Match(text);
-
-            if (match != null)
-                return new AudioBookTrack()
-                {
-                    Track = Convert.ToInt32(match.Groups[2].Value),
-                    FileName = match.Groups[1].Value,
-                    Title = match.Groups[4].Value
-                };
-
-
-            return null;
-        }
-
-        public static IEnumerable<AudioBookTrack> ParseMekMP3Page(string html)
-        {
-            var tracks = new List<AudioBookTrack>();
-
-            var document = new HtmlDocument();
-            document.Load(new StringReader(html));
-            foreach (var li in document.DocumentNode.SelectNodes("//li"))
-                try
-                {
-                    tracks.Add(MekConvert.ConvertAudioTrackFrom(li.InnerText));
-                }
-                catch
-                {
-                    //TODO: log parse error
-                }
-            return tracks;
-        }
-
-        public static IQueryable<Book> ParseMekBookResultPage(string pageContent)
-        {
-            var books = new List<Book>();
-
-            var document = new HtmlDocument();
-            document.Load(new StringReader(pageContent));
-            var docNode = document.DocumentNode;
-            foreach (var f in docNode.SelectNodes("//a[contains(@href,'Javascript')]"))
-            {
-                try
-                {
-                    var url = f.ParentNode.ParentNode.SelectSingleNode("span").FirstChild.InnerText;
-                    var catalog = CatalogResolver.Resolve(url);
-                    books.Add(new Book()
-                    {
-                        FullTitle = MekConvert.ClearText(f.InnerText),
-                        Id = catalog?.Id,
-                        UrlId = catalog?.UrlId
-                    });
-                }
-                catch (Exception ex)
-                {
-                    //TODO: log error
-                }
-            }
-            return books.AsQueryable();
-        }
     }
 }
